nullable gone

+++ Assignment-WebBank/Model/IndexModelProps.cs
@@ -8,7 +8,7 @@
         public string? City { get; set; }
         public int CustomerCount { get; set; }
         public int AccountCount { get; set; }
-        public int TotalBalance { get; set; }
+        public decimal TotalBalance { get; set; }
        
     }
 }
